deertogether/backend/requirements.txt
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.1
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
